requirements.txt
bs4==0.0.1
google==3.0.0
numpy==1.22.1
nltk==3.6.7
pdfplumber==0.6.0
tensorflow-cpu==2.7.0
spacy==3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz

Flask~=2.0.2
requests~=2.27.1
beautifulsoup4~=4.10.0